refactor: split routes/markets.py into resolution.py and comments.py
Closes #127

Split the 753-line routes/markets.py into three focused modules:
- routes/markets.py (310 lines) — CRUD: list, get, create, history
- routes/resolution.py (392 lines) — resolve, request-resolution,
  resolution-vote, committee-votes, resolution-votes
- routes/comments.py (107 lines) — get/create comments

New routers registered in routes/__init__.py and api.py.
All 206 existing tests pass with no behavior changes.

Rebased cleanly on main (including issue #115 tradeable-until-resolution,
structured logging, and split-storage changes).
